Change components folder structure
