+++ client/components/Home.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import Header from './Header';
-import Footer from './Footer';
+import Header from './commons/Header';
+import Footer from './commons/Footer';
 
 /**
  *
