+++ client/components/auth/LogoutButton.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
-import { logout } from '../actions/actionCreators';
+import { logout } from '../../actions/actionCreators';
 
 export const Logout = props => {
   const handleLogout = () => {
