+++ client/components/forms/EditCenterForm.jsx
@@ -3,10 +3,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
-import { fetchSingleCenter, updateCenter } from '../actions/actionCreators';
-import Header from './Header';
-import Footer from './Footer';
-import Loading from './Loading';
+import { fetchSingleCenter, updateCenter } from '../../actions/actionCreators';
+import Header from '../commons/Header';
+import Footer from '../commons/Footer';
+import Loading from '../commons/Loading';
 
 /**
  * Edits center details
