+++ client/components/presentationals/CenterDetailsComponent.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 
-import Loading from '../components/Loading';
+import Loading from '../commons/Loading';
 
 const CenterDetailsComponent = props => {
   const { selectedCenter, match, handleDelete, isFetching, handleModalOpen } = props;
