+++ client/components/presentationals/EventComponent.jsx
@@ -2,8 +2,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 
-import Event from '../components/Event';
-
 const EventComponent = ({ singleEvent, match }) => {
   const {
     centerLocation,
