+++ client/components/presentationals/EventDetailsComponent.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 
-import Loading from '../components/Loading';
+import Loading from '../commons/Loading';
 
 const EventDetailsComponent = ({
   currentEvent,
