+++ client/components/presentationals/EventListComponent.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 
-import Loading from '../components/Loading';
+import Loading from '../commons/Loading';
 import EventComponent from './EventComponent';
 
 const EventListComponent = ({ events, match, isFetching, error }) => {
