--- client/forms/EventForm.jsx
@@ -1,4 +1,0 @@
-// import React from 'react';
-// import { Link } from 'react-router-dom'
-
-// const
